[面試經驗] 取出可共用的上傳元件 (#1201)
title:	[面試經驗] 取出可共用的上傳元件
state:	OPEN
author:	peteranny
labels:	
assignees:	
reviewers:	barry800414 (Approved)
projects:	
milestone:	
number:	1201
url:	https://github.com/goodjoblife/GoodJobShare/pull/1201
additions:	174
deletions:	132
auto-merge:	disabled
--
#886 

## 這個 PR 是？ <!-- 必填 -->

面試經驗的上傳部分可以抽出，與其他表單共用


## 我應該如何手動測試？ <!-- 必填 -->

- [ ] http://localhost:3000/share 面試表單應該不受影響
